move buttons up a directory
