+++ frontend/src/components/SearchField/SearchField.tsx
@@ -1,6 +1,6 @@
 import { TextField } from '@mui/material';
 import { ChangeEvent, FormEvent, useState } from 'react';
-import SearchButton from '../Navigation/Desktop/Buttons/SearchButton/SearchButton';
+import SearchButton from '../Navigation/Desktop/Buttons/SearchButton';
 
 function SearchField(): JSX.Element {
   function search(event: FormEvent) {
